Moved map related field code to separate file
