+++ src/enums.h
@@ -2,6 +2,7 @@
 #ifndef ENUMS_H
 #define ENUMS_H
 
+#include <array>
 #include <cassert>
 #include <climits>
 #include <ostream>
@@ -355,6 +356,18 @@ inline bool operator<( const tripoint &a, const tripoint &b )
     }
     return false;
 }
+
+static const std::array<tripoint, 8> eight_horizontal_neighbors = { {
+        { -1, -1, 0 },
+        {  0, -1, 0 },
+        { +1, -1, 0 },
+        { -1,  0, 0 },
+        { +1,  0, 0 },
+        { -1, +1, 0 },
+        {  0, +1, 0 },
+        { +1, +1, 0 },
+    }
+};
 
 struct rectangle {
     point p_min;
