+++ src/field.h
@@ -133,15 +133,6 @@ extern field_id field_from_ident( const std::string &field_ident );
 bool field_type_dangerous( field_id id );
 
 /**
- * converts wind direction to list of co-ords to block neighbours to spread to.
- */
-std::tuple<maptile, maptile, maptile> get_wind_blockers( const int &winddirection,
-        const tripoint &pos );
-/**
- * converts xy of disallowed wind directions to map tiles.
- */
-
-/**
  * An active or passive effect existing on a tile.
  * Each effect can vary in intensity (density) and age (usually used as a time to live).
  */
